[Behat][Maintenance] Move pages to proper namespaces
