+++ tests/Behat/Context/Ui/Admin/ManagingInvoicesContext.php
@@ -6,9 +6,9 @@ namespace Tests\Sylius\InvoicingPlugin\Behat\Context\Ui\Admin;
 
 use Behat\Behat\Context\Context;
 use Sylius\Component\Core\Model\OrderInterface;
-use Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\IndexPageInterface;
-use Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\OrderShowPageInterface;
-use Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\ShowPageInterface;
+use Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\Invoice\IndexPageInterface;
+use Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\Invoice\ShowPageInterface;
+use Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\Order\ShowPageInterface as OrderShowPageInterface;
 use Webmozart\Assert\Assert;
 
 final class ManagingInvoicesContext implements Context
@@ -20,16 +20,16 @@ final class ManagingInvoicesContext implements Context
     private $showPage;
 
     /** @var OrderShowPageInterface */
-    private $orderPage;
+    private $orderShowPage;
 
     public function __construct(
         IndexPageInterface $indexPage,
         ShowPageInterface $showPage,
-        OrderShowPageInterface $orderPage
+        OrderShowPageInterface $orderShowPage
     ) {
         $this->indexPage = $indexPage;
         $this->showPage = $showPage;
-        $this->orderPage = $orderPage;
+        $this->orderShowPage = $orderShowPage;
     }
 
     /**
@@ -76,7 +76,7 @@ final class ManagingInvoicesContext implements Context
      */
     public function shouldSeeAnInvoiceRelatedToTheOrder(): void
     {
-        Assert::true($this->orderPage->hasRelatedInvoices(1));
+        Assert::true($this->orderShowPage->hasRelatedInvoices(1));
     }
 
     /**
@@ -84,7 +84,7 @@ final class ManagingInvoicesContext implements Context
      */
     public function clickOnFirstInvoiceId(): void
     {
-        $this->orderPage->clickOnFirstInvoiceId();
+        $this->orderShowPage->clickOnFirstInvoiceId();
     }
 
     /**
@@ -92,7 +92,7 @@ final class ManagingInvoicesContext implements Context
      */
     public function iClickOnFirstInvoiceDownloadButton(): void
     {
-        $this->orderPage->downloadFirstInvoice();
+        $this->orderShowPage->downloadFirstInvoice();
     }
 
     /**
@@ -163,6 +163,6 @@ final class ManagingInvoicesContext implements Context
      */
     public function thePdfFileForThisInvoiceShouldBeDownloadedSuccessfully(): void
     {
-        Assert::true($this->orderPage->isPdfFileDownloaded());
+        Assert::true($this->orderShowPage->isPdfFileDownloaded());
     }
 }
