+++ tests/Behat/Page/Admin/Order/ShowPageInterface.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Tests\Sylius\InvoicingPlugin\Behat\Page\Admin;
+namespace Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\Order;
 
 use Sylius\Behat\Page\SymfonyPageInterface;
 
-interface OrderShowPageInterface extends SymfonyPageInterface
+interface ShowPageInterface extends SymfonyPageInterface
 {
     public function hasRelatedInvoices(int $count): bool;
