+++ tests/Behat/Resources/services.xml
@@ -3,20 +3,21 @@
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <services>
         <service id="Tests\Sylius\InvoicingPlugin\Behat\Context\Ui\Admin\ManagingInvoicesContext" public="true">
-            <argument type="service" id="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\IndexPageInterface" />
-            <argument type="service" id="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\ShowPageInterface" />
-            <argument type="service" id="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\OrderShowPageInterface" />
+            <argument type="service" id="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\Invoice\IndexPageInterface" />
+            <argument type="service" id="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\Invoice\ShowPageInterface" />
+            <argument type="service" id="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\Order\ShowPageInterface" />
             <tag name="fob.context_service" />
         </service>
 
-        <service id="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\IndexPageInterface" class="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\IndexPage"
+        <service id="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\Invoice\IndexPageInterface" class="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\Invoice\IndexPage"
                  parent="sylius.behat.page.admin.crud.index">
             <argument type="string">sylius_invoicing_plugin_admin_invoice_index</argument>
         </service>
 
-        <service id="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\ShowPageInterface" class="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\ShowPage" parent="sylius.behat.symfony_page">
+        <service id="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\Invoice\ShowPageInterface" class="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\Invoice\ShowPage" parent="sylius.behat.symfony_page">
             <argument type="service" id="sylius.behat.table_accessor" />
         </service>
-        <service id="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\OrderShowPageInterface" class="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\OrderShowPage" parent="sylius.behat.symfony_page" />
+
+        <service id="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\Order\ShowPageInterface" class="Tests\Sylius\InvoicingPlugin\Behat\Page\Admin\Order\ShowPage" parent="sylius.behat.symfony_page" />
     </services>
 </container>
